fix(ui): normalize responsive dialog constraint types

+++ flutter_app/lib/shared/widgets/folder_browser_dialog.dart
@@ -128,8 +128,8 @@ class _FolderBrowserDialogState extends State<FolderBrowserDialog> {
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
       child: ConstrainedBox(
         constraints: BoxConstraints(
-          maxWidth: dialogWidth.clamp(280.0, 600.0),
-          maxHeight: dialogHeight.clamp(280.0, 500.0),
+          maxWidth: dialogWidth.clamp(280.0, 600.0).toDouble(),
+          maxHeight: dialogHeight.clamp(280.0, 500.0).toDouble(),
         ),
         child: Padding(
           padding: const EdgeInsets.all(DesignTokens.space4),
